Move run tbb pipeline function from cpp to hpp headers
